Move global app-specific code inside app directory
