+++ src/app/components/AppBottomBar.svelte
