+++ src/app/components/AppHeader.svelte
@@ -1,8 +1,9 @@
 <script>
 import { createEventDispatcher } from 'svelte';
 import dayjs from 'dayjs';
-import Button from '@components/Button.svelte';
+
 import { todos, removeDoneTimer } from '@features/todos/store';
+import Button from '@components/Button.svelte';
 
 $: canUndoRemove = $removeDoneTimer > 0;
 $: hasDoneTodos = $todos.some((todo) => todo.done);
