+++ src/app/stores/confirmation.js
