+++ src/index.js
@@ -1,9 +1,9 @@
-import App from './App.svelte';
+import App from '@app/App.svelte';
 import { initializeSentry } from '@lib/sentry';
 import { initializeThemes } from '@features/themes';
 import { initializeBackgrounds } from '@features/backgrounds';
 import { initializeKeyBindings } from '@features/shortcuts';
-import '@styles/index.css';
+import '@app/app.css';
 
 initializeSentry();
 initializeThemes();
